Refactor: move the crio config to the crio runtime
For some reason it was left with the generic CRI
